chore: retire google vertex AI
