+++ core/src/lib.rs
@@ -38,7 +38,6 @@ pub mod providers {
     }
     pub mod anthropic;
     pub mod google_ai_studio;
-    pub mod google_vertex_ai;
 }
 pub mod http {
     pub mod request;
