+++ front/pages/api/w/[wId]/providers/[pId]/models.ts
@@ -279,11 +279,6 @@ async function handler(
           }
           return;
 
-        case "google_vertex_ai":
-          return res.status(200).json({
-            models: [{ id: "gemini-pro" }],
-          });
-
         case "google_ai_studio":
           return res.status(200).json({
             models: [{ id: "gemini-pro" }, { id: "gemini-1.5-pro-latest" }],
