+++ types/src/front/provider.ts
@@ -15,7 +15,5 @@ export type CredentialsType = {
   SERP_API_KEY?: string;
   SERPER_API_KEY?: string;
   BROWSERLESS_API_KEY?: string;
-  GOOGLE_VERTEX_AI_SERVICE_ACCOUNT_JSON?: string;
-  GOOGLE_VERTEX_AI_ENDPOINT?: string;
   GOOGLE_AI_STUDIO_API_KEY?: string;
 };
